Updating Q03

+++ L11 Test/Test Preparation III/PT III/Q02 V2/Program.cs
@@ -21,9 +21,11 @@ public class Program
                     break;
 
                 case "sort":
+                    SortSubArray(array, inputTokens);
                     break;
 
                 case "rollLeft":
+                    CommandRollLeft(array, inputTokens);
                     break;
 
                 case "rollRight":
@@ -32,6 +34,41 @@ public class Program
 
             input = Console.ReadLine();
         }
+
+        string outPut = string.Join(", ", array);
+        Console.WriteLine($"[{outPut}]");
+    }
+
+    public static void CommandRollLeft(List<string> array, List<string> inputTokens)
+    {
+        int shiftBy = int.Parse(inputTokens[1]);
+
+        for (int index = array.Count() - 1; index >= 0; index--)
+        {
+            var currentString = array[index];
+
+
+        }
+    }
+
+    public static void SortSubArray(List<string> array, List<string> inputTokens)
+    {
+        int startIndex = int.Parse(inputTokens[2]);
+        int count = int.Parse(inputTokens[4]);
+
+        bool validIndexs = IndexValidator(array, startIndex) && IndexValidator(array, startIndex + count);
+        if (!validIndexs)
+        {
+            Console.WriteLine("Invalid input parameters.");
+            return;
+        }
+
+        var newSubArray = array.GetRange(startIndex, count).Select(x => x).OrderBy(x => x).ToList();
+
+        array.RemoveRange(startIndex, count);
+        array.InsertRange(startIndex, newSubArray);
+
+        Console.WriteLine(string.Join(" ", array));
     }
 
     public static void ReverseSubArray(List<string> array, List<string> inputTokens)
@@ -40,13 +77,16 @@ public class Program
         int count = int.Parse(inputTokens[4]);
 
         bool validIndexs = IndexValidator(array, startIndex) && IndexValidator(array, startIndex + count);
+        if (!validIndexs)
+        {
+            Console.WriteLine("Invalid input parameters.");
+            return;
+        }
 
         var newSubArray = array.GetRange(startIndex, count).Select(x => x).Reverse().ToList();
 
         array.RemoveRange(startIndex, count);
         array.InsertRange(startIndex, newSubArray);
-
-        Console.WriteLine(string.Join(" ", array));
     }
 
     public static bool IndexValidator(List<string> array, int currentIndex)
